backend/requirements.txt
alembic==1.13.0
annotated-types>=0.7.0
anthropic==0.8.1
anyio>=3.7.1
asyncpg==0.29.0
backoff==2.2.1
bcrypt>=4.0.1
boto3>=1.42.49
botocore>=1.42.49
certifi>=2026.1.4
cffi>=2.0.0
charset-normalizer>=3.4.4
chromadb>=0.5.23
click>=8.3.1
coverage>=7.13.4
cryptography>=46.0.5
distro>=1.9.0
ecdsa>=0.19.1
fastapi>=0.104.1
filelock>=3.24.0
flatbuffers>=25.12.19
fsspec>=2026.2.0
google-api-core>=2.29.0
google-auth>=2.48.0
google-cloud-aiplatform>=1.115.0
google-cloud-bigquery>=3.13.0
google-cloud-core>=2.5.0
google-cloud-resource-manager>=1.16.0
google-cloud-storage>=2.19.0
google-crc32c>=1.8.0
google-genai>=1.2.0
google-generativeai>=0.8.6
google-resumable-media>=2.8.0
googleapis-common-protos>=1.72.0
greenlet>=3.0.1
grpcio>=1.78.0
grpcio-status>=1.62.3
h11>=0.16.0
httpcore>=1.0.9
httptools>=0.7.1
httpx>=0.25.2
idna>=3.11
Jinja2>=3.1.6
jsonschema>=4.26.0
kubernetes>=35.0.0
langchain-core>=1.2.7
langgraph>=1.0.6
langgraph-checkpoint>=4.0.0
langgraph-prebuilt>=1.0.6
langgraph-sdk>=0.3.3
langsmith>=0.6.4
Mako>=1.3.10
MarkupSafe>=3.0.3
mmh3>=5.2.0
numpy>=1.26.4
oauthlib>=3.3.1
onnxruntime>=1.24.1
openai>=1.3.7
opentelemetry-api>=1.27.0
opentelemetry-exporter-otlp-proto-common>=1.27.0
opentelemetry-exporter-otlp-proto-grpc>=1.27.0
opentelemetry-instrumentation>=0.48b0
opentelemetry-instrumentation-asgi>=0.48b0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-proto>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-semantic-conventions>=0.48b0
opentelemetry-util-http>=0.48b0
orjson>=3.11.0
packaging>=25.0
pandas>=2.0.0
passlib>=1.7.4
plotly>=5.0.0
posthog>=5.4.0
proto-plus>=1.27.1
protobuf>=4.25.8
psycopg2-binary>=2.9.9
pydantic>=2.7.4
pydantic-settings>=2.1.0
PyAthena>=3.2.0
PyYAML>=6.0.2
redis>=5.0.1
requests>=2.32.5
requests-oauthlib>=2.0.0
SQLAlchemy>=2.0.23
starlette>=0.27.0
tenacity>=9.1.2
tokenizers>=0.20.3
tqdm>=4.67.3
typing_extensions>=4.15.0
urllib3>=2.6.3
uvicorn>=0.24.0
uvloop>=0.21.0
vanna>=0.7.9
watchfiles>=1.1.0
websocket-client>=1.8.0
websockets>=14.2
wrapt>=1.17.2
python-dotenv>=1.0.0
python-jose>=3.3.0
python-multipart>=0.0.6
python-dateutil>=2.9.0
